Added randomForest and knn models
I was having a bug when running knn with scikit-learn 1.3.0, so I uninstalled and reinstalled with 1.2.2. This fixed the bug but caused a bunch of "you're using an outdated version of scikit-learn" warnings in the terminal, so I added a warning suppressor to the code. Try running with scikit-learn 1.3.0 first, if that doesn't work, run with 1.2.2.